Page break: Fix small logic error
Don't mark a frame that as pushed (will prevent further page-break
checks), if the break actually occured on a frame before it in the tree;
only the latter one should not be checked again, the original might find
itself in a different situation on the next page and might need to be
checked again.

The logic might actually be correct, if there was some logic to revert
already decided page breaks, in case no other frame alllowing breaks was
found, but for now this should be an improvement.